refactor: rename NoteMultiSelect to NoteSelect (add single mode) and TagMultiSelect to TagSelect
